# book-network/src/main/java/com/mpn/book/book/BookResponse.java
package com.mpn.book.book;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class BookResponse {

    private Integer id;
    private String title;
    private String authorName;
    private String isbn;
    private String synopsis;
    private String owner;
    private byte[] cover;
    private Double rate;
    private boolean archived;
    private boolean shareable;

}