# book-network/src/main/java/com/mpn/book/feedback/FeedbackResponse.java
package com.mpn.book.feedback;

import lombok.*;

@Getter
@Setter
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class FeedbackResponse {

    private Double note;
    private String comment;
    private boolean ownFeedBack;

}